+++ src/FsCodec.SystemTextJson/Interop.fs
@@ -42,10 +42,8 @@ type InteropExtensions =
         else JsonSerializer.Deserialize(System.ReadOnlySpan.op_Implicit x)
     static member private MapTo(x: JsonElement) : byte[] =
         if x.ValueKind = JsonValueKind.Undefined then null
-        else JsonSerializer.SerializeToUtf8Bytes(x, options = InteropExtensions.NoOverEscapingOptions)
-    // Avoid introduction of HTML escaping for things like quotes etc (as standard Options.Create() profile does)
-    static member private NoOverEscapingOptions =
-        System.Text.Json.JsonSerializerOptions(Encoder = System.Text.Encodings.Web.JavaScriptEncoder.UnsafeRelaxedJsonEscaping)
+        // Avoid introduction of HTML escaping for things like quotes etc (Options.Default uses Options.Create(), which defaults to unsafeRelaxedJsonEscaping=true)
+        else JsonSerializer.SerializeToUtf8Bytes(x, options = Options.Default)
 
     [<Extension>]
     static member ToByteArrayCodec<'Event, 'Context>(native : FsCodec.IEventCodec<'Event, JsonElement, 'Context>)